requirements.txt
requests
pillow
gradio
